// frontend/src/client/models/ArticleRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ArticleRead = {
    id: number;
    title: string;
    content: string;
    author: string;
    published_at?: (string | null);
    is_published: boolean;
};